requirements.txt
flatlib==0.2.5
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
pyyaml
